Docs: Help people upgrade to 4.0.0
The changelog should clearly state whether 4.0.0 is a change to the public API of `browserify-rails` or not.

The phrase "change to new API usage" seems to refer to the new sprockets version constraint, but could be misinterpreted to mean that the public API of `browserify-rails` has changed. Maybe it is more clear to simply say "depend on sprockets >= 3.6.0"?