Update version of browserify in readme
This is a pull request, but it's also a question, ie. does browserify-rails support the latest browserify (currently 14.1.0)?

I made a PR just in case you wanted to have the latest version number in the readme.

I changed the tilde to a caret because that's what my team uses, but again, maybe not what you want in the readme.

Thanks!