Re-open the output file before reading the results from browserify
This issue raised its head with changes to browserify for 14.1.0. This
version introduces a two-step process for generating results in which
results are written to a new file and then copied into the path
designated in the options. This means that the file handle
browserify-rails has from `Tempfile` is no longer valid and is always
empty.

This is further complicated in that it is only a situation when running
browserify directly and not browserifyinc. I noticed there weren't any
tests disabling browserifyinc so hit two issues with one test.